requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
firebase-admin==6.4.0
google-cloud-storage==2.14.0
qrcode[pil]==7.4.2
requests==2.31.0
Jinja2==3.1.3
python-multipart==0.0.9
pydantic==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4


